# ARROW-4503: [C#] Eliminate allocations in ArrowStreamReader when reading from a Stream