--- a/csharp/src/Apache.Arrow/ArrowBuffer.Builder.cs
+++ b/csharp/src/Apache.Arrow/ArrowBuffer.Builder.cs
@@ -109,20 +109,19 @@ namespace Apache.Arrow
                 return this;
             }
 
-            public ArrowBuffer Build(MemoryPool pool = default)
+            public ArrowBuffer Build(MemoryAllocator allocator = default)
             {
-                int length;
-                checked
+                int length = checked((int)BitUtility.RoundUpToMultipleOf64(_buffer.Length));
+
+                var memoryAllocator = allocator ?? MemoryAllocator.Default.Value;
+                var memoryOwner = memoryAllocator.Allocate(length);
+
+                if (memoryOwner != null)
                 {
-                    length = (int)BitUtility.RoundUpToMultipleOf64(_buffer.Length);
+                    Memory.CopyTo(memoryOwner.Memory);
                 }
 
-                var memoryPool = pool ?? MemoryPool.Default.Value;
-                var memory = memoryPool.Allocate(length);
-
-                Memory.CopyTo(memory);
-
-                return new ArrowBuffer(memory);
+                return new ArrowBuffer(memoryOwner);
             }
 
             private Span<T> EnsureCapacity(int len)