--- a/csharp/src/Apache.Arrow/Ipc/ArrowFileReader.cs
+++ b/csharp/src/Apache.Arrow/Ipc/ArrowFileReader.cs
@@ -35,18 +35,18 @@ namespace Apache.Arrow.Ipc
         {
         }
 
-        public ArrowFileReader(Stream stream, MemoryPool memoryPool)
-            : this(stream, memoryPool, leaveOpen: false)
+        public ArrowFileReader(Stream stream, MemoryAllocator allocator)
+            : this(stream, allocator, leaveOpen: false)
         {
         }
 
         public ArrowFileReader(Stream stream, bool leaveOpen)
-            : this(stream, memoryPool: null, leaveOpen)
+            : this(stream, allocator: null, leaveOpen)
         {
         }
 
-        public ArrowFileReader(Stream stream, MemoryPool memoryPool, bool leaveOpen)
-            : base(new ArrowFileReaderImplementation(stream, memoryPool, leaveOpen))
+        public ArrowFileReader(Stream stream, MemoryAllocator allocator, bool leaveOpen)
+            : base(new ArrowFileReaderImplementation(stream, allocator, leaveOpen))
         {
         }