… csharp/src/Apache.Arrow/Memory/MemoryAllocator.cs
… csharp/src/Apache.Arrow/Memory/NativeMemoryAllocator.cs
… csharp/src/Apache.Arrow/Memory/NativeMemoryManager.cs
… csharp/src/Apache.Arrow/RecordBatch.cs
… csharp/test/Apache.Arrow.Benchmarks/ArrowReaderBenchmark.cs
… csharp/test/Apache.Arrow.Tests/ArrowBufferBuilderTests.cs
… csharp/test/Apache.Arrow.Tests/ArrowBufferTests.cs
… csharp/test/Apache.Arrow.Tests/ArrowFileReaderTests.cs
… csharp/test/Apache.Arrow.Tests/ArrowStreamReaderTests.cs
… csharp/test/Apache.Arrow.Tests/Fixtures/DefaultMemoryAllocatorFixture.cs
… csharp/test/Apache.Arrow.Tests/TestMemoryAllocator.cs
… csharp/test/Apache.Arrow.Tests/TestMemoryPool.cs