@eerhardt 2019-03-15
Here are some preliminary perf results of this change:

|                      Method |      Mean |     Error |    Gen 0/1k Op | Allocated Memory/Op |
|---------------------------- |----------:|----------:|----------:|--------------------:|
| **BEFORE** ArrowReaderWithMemoryStream | 81.417 ms | 1.6083 ms | 3000.0000 |         84000.73 KB |
| **AFTER** ArrowReaderWithMemoryStream | 29.431 ms | 0.5826 ms |             - |             6.28 KB |

The reason `Allocated Memory/Op` drops to almost 0 is because we are allocating from the default `MemoryPool`, which pulls from native memory, which isn't tracked by BenchmarkDotNet.