@chutchinson 2019-03-16
@eerhardt Is there a way to report the native memory to the benchmark? Seems like it would be valuable for benchmarks, especially considering the current allocator is naive and probably even broken and/or not thread-safe. I expect it will be replaced or supplemented with Jemalloc.NET or some other more reliable allocation mechanism.