@eerhardt 2019-03-16
> Is there a way to report the native memory to the benchmark?

I'm not sure, but maybe @adamsitnik might be able to answer.

On a related topic, @chutchinson, is there a specific reason why the default MemoryPool uses native memory? I guess my assumption was that the default should just be regular `byte[]` instances, which could be garbage collected. I also noticed that there doesn't seem to be a way to return memory back to the pool to be reused. Is that intentional?