@stephentoub 2019-03-17
> ArrowReaderWithMemoryStream

Do we expect this to be a common case? If so, have you considered extracting the byte[] underlying the MemoryStream and using the memory-based reader instead of stream-based? Your need to read the header info at least presumably to know how to slice it, but that should then should then allow that case to be zero-copy, too.