@chutchinson 2019-03-17
> On a related topic, @chutchinson, is there a specific reason why the default MemoryPool uses native memory? I guess my assumption was that the default should just be regular `byte[]` instances, which could be garbage collected. I also noticed that there doesn't seem to be a way to return memory back to the pool to be reused. Is that intentional?

The default memory pool allocates native memory because the allocated pointer can be adjusted to satisfy the Arrow memory alignment rule in a cross-platform way without GC pinning and the associated complexity. Admittedly I don't have a sufficiently deep understanding of the CLR (especially CoreCLR) to imagine a better way. 

There needs to be more thought and care put into the design of the memory allocation API and implementation. There is no mechanism for returning allocated memory; at the moment it is garbage collected as appropriate via the MemoryManager<T> API.