@stephentoub 2019-03-17
> There is no mechanism for returning allocated memory; at the moment it is garbage collected as appropriate via the MemoryManager API.

Given this, I'd recommend not using this pool in this PR.  It's going to make things significantly less efficient.