@adamsitnik 2019-03-19
> Is there a way to report the native memory to the benchmark?