@eerhardt 2019-05-03
OK - I finally got back to this PR and it is ready for re-review. I believe I have addressed all feedback and concerns. Please let me know.

Here are some updated numbers with the new benchmarks added:

## Before
|                      Method |   Count |         Mean |        Error |      StdDev | Gen 0/1k Op | Allocated Memory/Op |
|---------------------------- |-------- |-------------:|-------------:|------------:|------------:|--------------------:|
| ArrowReaderWithMemoryStream |   10000 |    639.69 us |    25.797 us |    71.91 us |   3000.0000 |           947.22 KB |
|       ArrowReaderWithMemory |   10000 |     59.98 us |     6.349 us |    17.70 us |           - |             4.48 KB |
| ArrowReaderWithMemoryStream | 1000000 | 78,008.70 us | 1,536.342 us | 1,437.10 us |   3000.0000 |         83999.47 KB |
|       ArrowReaderWithMemory | 1000000 |  2,675.14 us |   119.735 us |   333.77 us |           - |             4.48 KB |

## After
|                                    Method |   Count |         Mean |      Error |    StdDev | Gen 0/1k Op |  Allocated Memory/Op |
|------------------------------------------ |-------- |-------------:|-----------:|----------:|------------:|--------------------:|
|               ArrowReaderWithMemoryStream |   10000 |    185.59 us |  22.262 us |  65.29 us |   1000.0000 |              5.2 KB |
| ArrowReaderWithMemoryStream_ManagedMemory |   10000 |    110.95 us |  11.195 us |  32.12 us |           - |             5.19 KB |
|                     ArrowReaderWithMemory |   10000 |     55.50 us |   5.738 us |  16.37 us |           - |              4.7 KB |
|               ArrowReaderWithMemoryStream | 1000000 | 34,292.18 us | 582.247 us | 544.63 us |   1000.0000 |              5.2 KB |
| ArrowReaderWithMemoryStream_ManagedMemory | 1000000 | 27,583.96 us | 600.099 us | 616.26 us |           - |          41997.4 KB |
|                     ArrowReaderWithMemory | 1000000 |  2,897.50 us | 157.003 us | 462.93 us |           - |              4.7 KB |

Notes on the benchmarks:
1. I couldn't do the `_ManagedMemory` tests before this change, since that API didn't exist.
2. Notice the gains on `ArrowReaderWithMemoryStream_ManagedMemory` with `Count=10,000`. It is a 40% gain vs. the default case. The reasoning is because the `_ManagedMemory` uses `MemoryPool<byte>` to allocate managed memory. In the case of `10,000` records (the default in Spark), we are able to pool the memory buffers and don't need to allocate a new buffer each iteration.