@eerhardt 2019-05-06
> I also wonder if we could somehow force the users to Dispose the Batch properly?

It is `IDisposable`, I would assume any analyzer that enforced `IDisposable` should work. e.g. https://docs.microsoft.com/en-us/visualstudio/code-quality/ca2213-disposable-fields-should-be-disposed